requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
easyocr==1.7.2
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.1
opencv-python-headless==4.13.0.90
packaging==25.0
pillow==12.1.0
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.6.0
pytesseract==0.3.13
python-bidi==0.6.7
python-docx==1.2.0
python-multipart==0.0.21
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.9.0
shapely==2.1.2
starlette==0.50.0
sympy==1.14.0
tifffile==2026.1.14
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
